+++ api-controllers/DashboardController.js
@@ -5,11 +5,11 @@ const _ = require('lodash');
 const yaml = require('js-yaml');
 const Promise = require('bluebird');
 const Session = require('express-session').Session;
-const config = require('../config');
-const errors = require('../errors');
-const logger = require('../logger');
-const catalog = require('../models').catalog;
-const utils = require('../utils');
+const config = require('../common/config');
+const errors = require('../common/errors');
+const logger = require('../common/logger');
+const catalog = require('../broker/lib/models').catalog;
+const utils = require('../broker/lib/utils');
 const FabrikBaseController = require('./FabrikBaseController');
 const Forbidden = errors.Forbidden;
 const ContinueWithNext = errors.ContinueWithNext;
