+++ api-controllers/ServiceFabrikAdminController.js
@@ -3,20 +3,20 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 var moment = require('moment-timezone');
-const catalog = require('../models/catalog');
-const errors = require('../errors');
-const logger = require('../logger');
-const config = require('../config');
-const NetworkSegmentIndex = require('../bosh/NetworkSegmentIndex');
-const backupStore = require('../iaas').backupStore;
+const catalog = require('../broker/lib/models/catalog');
+const errors = require('../common/errors');
+const logger = require('../common/logger');
+const config = require('../common/config');
+const NetworkSegmentIndex = require('../broker/lib/bosh/NetworkSegmentIndex');
+const backupStore = require('../broker/lib/iaas').backupStore;
 const FabrikBaseController = require('./FabrikBaseController');
-const utils = require('../utils');
-const fabrik = require('../fabrik');
-const bosh = require('../bosh');
-const ScheduleManager = require('../jobs');
-const BackupReportManager = require('../reports');
-const CONST = require('../constants');
-const maintenanceManager = require('../maintenance').maintenanceManager;
+const utils = require('../broker/lib/utils');
+const fabrik = require('../broker/lib/fabrik');
+const bosh = require('../broker/lib/bosh');
+const ScheduleManager = require('../broker/lib/jobs');
+const BackupReportManager = require('../broker/lib/reports');
+const CONST = require('../common/constants');
+const maintenanceManager = require('../broker/lib/maintenance').maintenanceManager;
 const Conflict = errors.Conflict;
 const Forbidden = errors.Forbidden;
 const BadRequest = errors.BadRequest;
