+++ api-controllers/ServiceFabrikApiController.js
@@ -3,19 +3,19 @@
 const _ = require('lodash');
 const assert = require('assert');
 const Promise = require('bluebird');
-const jwt = require('../jwt');
-const logger = require('../logger');
-const backupStore = require('../iaas').backupStore;
+const jwt = require('../broker/lib/jwt');
+const logger = require('../common/logger');
+const backupStore = require('../broker/lib/iaas').backupStore;
 const filename = backupStore.filename;
-const eventmesh = require('../../../eventmesh');
+const eventmesh = require('../eventmesh');
 const lockManager = eventmesh.lockManager;
-const errors = require('../errors');
+const errors = require('../common/errors');
 const FabrikBaseController = require('./FabrikBaseController');
 const Unauthorized = errors.Unauthorized;
 const NotFound = errors.NotFound;
 const Gone = errors.Gone;
 const Timeout = errors.Timeout;
-const cf = require('../cf');
+const cf = require('../broker/lib/cf');
 const Forbidden = errors.Forbidden;
 const BadRequest = errors.BadRequest;
 const Conflict = errors.Conflict;
@@ -25,12 +25,12 @@ const JsonWebTokenError = jwt.JsonWebTokenError;
 const ContinueWithNext = errors.ContinueWithNext;
 const InternalServerError = errors.InternalServerError;
 const DeploymentAlreadyLocked = errors.DeploymentAlreadyLocked;
-const ScheduleManager = require('../jobs');
-const config = require('../config');
-const CONST = require('../constants');
-const catalog = require('../models').catalog;
-const utils = require('../utils');
-const docker = config.enable_swarm_manager ? require('../docker') : undefined;
+const ScheduleManager = require('../broker/lib/jobs');
+const config = require('../common/config');
+const CONST = require('../common/constants');
+const catalog = require('../broker/lib/models').catalog;
+const utils = require('../broker/lib/utils');
+const docker = config.enable_swarm_manager ? require('../broker/lib/docker') : undefined;
 
 const CloudControllerError = {
   NotAuthorized: err => {
