+++ broker/ReportServer.js
@@ -1,8 +1,7 @@
 'use strict';
 
 console.log('Starting Service Fabrik Report App...');
-const lib = require('./lib');
-const routes = lib.routes;
+const routes = require('../api-controllers/routes');
 const HttpServer = require('../common/HttpServer');
 const ExpressApp = require('../common/ExpressApp');
