+++ broker/server.js
@@ -2,7 +2,7 @@
 
 console.log('Starting Service Fabrik...');
 const lib = require('./lib');
-const routes = lib.routes;
+const routes = require('../api-controllers/routes');
 const HttpServer = require('../common/HttpServer');
 const ExpressApp = require('../common/ExpressApp');
 const config = lib.config;
