+++ test/test_broker/fabrik.OOBBackupManager.spec.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const BoshDirectorClient = require('../../broker/lib/bosh/BoshDirectorClient');
 const BackupStore = require('../../broker/lib/iaas/BackupStore');
 const Agent = require('../../broker/lib/fabrik/Agent');
-const FabrikBaseController = require('../../broker/lib/controllers/FabrikBaseController');
+const FabrikBaseController = require('../../api-controllers/FabrikBaseController');
 const CONST = require('../../broker/lib/constants');
 const OobBackupManager = require('../../broker/lib/fabrik/OobBackupManager');
 const bosh = require('../../broker/lib/bosh');
