+++ test/test_broker/support/apps.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const _ = require('lodash');
-const lib = require('../../../broker/lib');
-const routes = lib.routes;
+//const lib = require('../../../broker/lib');
+const routes = require('../../../api-controllers/routes');
 const ExpressApp = require('../../../common/ExpressApp');
 
 // internal app
